+++ docs/HISTORY.rst
@@ -4,18 +4,19 @@ Changelog
 3.1.0 (2015-04-20)
 ------------------
 
-* Added the ``--verbose`` (``-v``) flag
+* Added the :option:`--verbose` (``-v``) flag
 * Following the split between getting the list of groups, and
   sending a digest
+* Moved the documentation to Read the Docs
 
 3.0.0 (2015-02-26)
 ------------------
 
-* Renaming the product `gs.group.messages.digest.send`_ from
+* Renaming the product `gs.group.messages.topic.digest.send`_ from
   ``gs.group.messages.senddigest``
 
 .. _gs.group.messages.digest.send:
-   https://github.com/groupserver/gs.group.messages.digest.send
+   https://github.com/groupserver/gs.group.messages.topic.digest.send
 
 2.1.0 (2014-03-26)
 ------------------